requirements.txt
#httpx~=0.28.1
#dotenv~=0.9.9
python-dotenv~=1.0.1
openai~=1.65.4
pydantic~=2.10.6
python-pptx~=1.0.2
